requirements.txt
appnope==0.1.0
backcall==0.1.0
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.3.0
dlib==19.10.0
enum34==1.1.6
face-recognition==1.2.2
face-recognition-models==0.3.0
Flask==1.0.2
Flask-Cors==3.0.4
html5lib==0.9999999
idna==2.6
imutils==0.4.3
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jupyter-client==5.2.3
jupyter-core==4.4.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
numpy==1.14.3
opencv-python==3.4.0.12
pandas==0.20.3
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
pyzmq==17.0.0
requests==2.18.4
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
gunicorn==19.7.1
gevent==1.2.2